MB-27243: log that we've received a request to delete a bucket
When deleting buckets we don't log prior to actually attempting the
deletion. It's a significant enough operation that adding
a debug log that we've received a request and are about to delete
seems reasonable. And it will make it simpler to know whether or
not deletion is hanging or if the request wasn't actually received.

Change-Id: I10f048d6ed6eeae1792940df7741f201893e6c9a
Reviewed-on: http://review.couchbase.org/86877
Tested-by: Dave Finlay <dave.finlay@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Poonam Dhavale <poonam@couchbase.com>