[async] Allow initiating adoption during async's initialization.
That helps avoid races where the prospective adopter dies before
adopting an async, so the async keeps running.

Note that care needs to be taken to ensure that the adopter is still
alive when the async tries to register with it. But that shouldn't be
an issue most of the times because the adopter is expected to use the
result produced by the adoptee. In order to get the said result, it
needs to wait on the adoptee anyway. So under normal circumstances it
should outlive the adopted async.

Change-Id: I2d3de7ab42e61016e21d101841d61022045b1cd6
Reviewed-on: http://review.couchbase.org/88112
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>