MB-26600: Add rest endpoints and config elements for logRedaction
Add a new log_redaction_default_cfg parameter to ns_config_default.
This contains
redact_level: can be set to "none"(default) or "partial"

Add GET and POST endpoints to read and commit values for these
parameters from/to config respectively.

Currently, this parameter will only be consumed by the UI.
On the log collection page the UI will autofill the cluster setting
for redact level as the default and the user will have the option
to override this setting if they so choose.

Change-Id: I2521055f4e958d61ed002830d1cc5d49d496b0bf
Reviewed-on: http://review.couchbase.org/87717
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Tested-by: Artem Stemkovski <artem@couchbase.com>