[gen_server2] Make it simple to wait for a condition.
Before running a function.

Another common pattern is handling client request, but only once
certain condition becomes satisfied. With this change this can be
achieved quite simply as follows:

  handle_call(do_something, From, State) ->
      gen_server2:conditional(fun check_condition/1,
                              fun (NewState) ->
                                  gen_server2:reply(From, satisfied),
                                  {noreply, update_state(NewState)}
                              end,
                              5000,
                              fun (NewState) ->
                                  gen_server2:reply(From, timeout),
                                  {noreply, update_state_on_timeout(NewState)}
                              end),
      {noreply, State};

Note, that the condition predicate is called whenever the server state
might have been updated. So it's quite simplistic in this
respect. Consequently, care needs to be taken to make the predicate as
cheap as possible.

Change-Id: I81210ece8a7b9ffe8a4314ce850f6182a31b6e5b
Reviewed-on: http://review.couchbase.org/88120
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>