[async] Reraise exceptions in executor process.
If the body function raises an exception of some sort, it would be
properly propagated to the caller. But the executor process would
terminate normally no matter what. So if the body linked itself to
some process in one way or another, the linked process would survive
the death of the async, which is obviously undesirable. With this
commit the executor process will re-raise the exception, so linked
processes will get terminated properly. For the caller of the async
nothing changes.

Change-Id: Ie7ba442dcd2bc3f6f7aee9eed633a7ed62822f43
Reviewed-on: http://review.couchbase.org/88339
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>