[async] Unwrap exceptions for 'perform' asyncs.
That way they behave more like regular processes. That is, if the body
function crashes, you'll see the same termination reason as if you
just used spawn_monitor, with the bonus of not having to take extra
care to ensure that child terminates if the parent dies.

Change-Id: I8887ffb22b28c0ceac60910393e0d01dab23c8ea
Reviewed-on: http://review.couchbase.org/88340
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>