Delay leader_lease_acquirer start till cluster is vulcan.
Technically, the process is started even before that, but it's not
doing anything useful. Otherwise, it would keep complaining about
failures to acquire leases from old nodes.

Note that this also mean that right after transition to vulcan, the
acquirer will have to acquire fresh leases from all nodes. This might
be suboptimal, but given that this happens only once, I'm willing to
tolerate that.

Change-Id: I54619ca1692b634ac3589939028a600fb1a1744a
Reviewed-on: http://review.couchbase.org/89219
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>