Handle not_a_leader error when forwarding whereis_name request.
The end behavior is unchanged: 'undefined' is returned to the
caller. The difference is in the verbosity of logging. Without special
handling, for each failed attempt a crash report will be logged. Since
it's not entirely unusual for a short period of time to have a node
be mistaken about the leader, I decided to handle this specially.

Change-Id: I717f68fb43a3cdaf076c1e85cda92dee86b541e8
Reviewed-on: http://review.couchbase.org/89250
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>