[leader_activities] Use only active nodes in default quorums.
I don't know how I missed it, but that was always my intention. The
idea is that we don't pass node additions through the orchestrator. I
didn't want to change that. So the nodes that are either inactive
because they are failed over or just added do not participate in
quorums. Then it's the responsibility of whoever changes node statuses
to active (rebalance and recovery) to have the majority of future
quorum. Similarly, it's the responsibility of those who deactivate
nodes (failover) to hold the majority of current the quorum to do so.

Change-Id: I71b7a86040c748f433c60c3712dfcc7ef663fcd3
Reviewed-on: http://review.couchbase.org/89424
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Poonam Dhavale <poonam@couchbase.com>