Lower wait_for_agents timeout for service failover.
If other non-failed-over nodes are down, then the whole failover will
fail (yet the node *will* get deactivated). The failure is most likely
to occur after one of the timeouts is hit. And wait_for_agents' one is
the most likely to get hit, since it's the first one. So lowering it
to improve user experience.

Change-Id: Ief1bf7bf9fb70f7a086db5480b9c5ab3fea4bb21
Reviewed-on: http://review.couchbase.org/89427
Reviewed-by: Poonam Dhavale <poonam@couchbase.com>
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>