[leader_activities] Allow inheriting options of parent activity.
Only works if an option in interest is not explicitly set by the child
activity. In addition, it's only supported for the activities started
by run_activity/start_activity functions. The reason register_process
is not supported is because we can't easily detect parent/child
relationship there. But essentially register_process should only be
used for the "root" activity anyway.

Change-Id: Ib1cebd2ecbb2f220b152494e80740653e6d9ca6b
Reviewed-on: http://review.couchbase.org/89580
Reviewed-by: Poonam Dhavale <poonam@couchbase.com>
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>